Fixed for non ssl connection
Shell in a box is working right.  We failed to understand what it is doing.  It forces ssl. So if you are in a non ssl session in omv the shell will not open because it knows the connections is not secure.  I added a shell web client button.  When you click on the button it will open up a new client that will redirect to ssl.  Users approve the cert then can use shell in the web client page even though their omv session is not secure.  The shell session will be secure.